+++ crates/oas3/src/spec/parameter.rs
@@ -1,8 +1,8 @@
 use serde::{Deserialize, Serialize};
 
 use super::{
-    spec_extensions, Example, FromRef, MediaType, ObjectOrReference, ObjectSchema, Ref, RefError,
-    RefType, Spec,
+    spec_extensions, Example, FromRef, MediaType, ObjectOrReference, Ref, RefError, RefType,
+    Schema, Spec,
 };
 use crate::Map;
 
@@ -185,7 +185,7 @@ pub struct Parameter {
     ///
     /// A parameter MUST contain either a schema property, or a content property, but not both.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub schema: Option<ObjectOrReference<ObjectSchema>>,
+    pub schema: Option<Schema>,
 
     /// Example of the parameter's potential value.
     ///