+++ crates/oas3/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
+- Support boolean schemas in schema-bearing fields including `spec::Components::schemas`,
+  `spec::MediaType::schema`, `spec::Parameter::schema`, `spec::Header::schema`, and
+  `spec::ObjectSchema::{all_of, any_of, one_of, prefix_items, properties}`.
+- Add `spec::Schema::resolve()` method.
 - The type of the `spec::Operation::callbacks` field is now `BTreeMap<String, ObjectOrReference<Callback>>`.
 - Add `spec::Operation::callbacks()` method.
+- Implement `Copy` for `spec::BooleanSchema`.
+- Remove implementation of `FromRef` for `spec::ObjectSchema`.
 - Minimum supported Rust version (MSRV) is now 1.87.
 
 ## 0.21.0